% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devoutverse.R
\docType{package}
\name{devoutverse}
\alias{devoutverse}
\title{devoutverse: A collection of bespoke, non-standard devices for RStats}
\description{
devoutverse: A collection of bespoke, non-standard devices for RStats
}
